added explicit cast

+++ src/types/coordinate.c
@@ -6,7 +6,7 @@
 /*   By: awillems <awillems@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/19 14:50:02 by awillems          #+#    #+#             */
-/*   Updated: 2022/12/21 11:33:51 by awillems         ###   ########.fr       */
+/*   Updated: 2022/12/21 11:41:30 by awillems         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,14 +30,14 @@ t_coord_f64	init_f64(void)
 	return (res);
 }
 
-t_coord_f64	set_f64(float x, float y)
+t_coord_f64	set_f64(double x, double y)
 {
 	return ((t_coord_f64){x, y});
 }
 
 t_coord_i32	set_i32(float x, float y)
 {
-	return ((t_coord_i32){x, y});
+	return ((t_coord_i32){(int) x, (int) y});
 }
 
 t_coord_f64	add_f64(t_coord_f64 a, t_coord_f64 b)
